Improvement: Refactor markdown list code
